+++ contracts/external/compound/CEther.sol
@@ -3,11 +3,18 @@ pragma solidity ^0.8.0;
 
 interface CEther {
     function mint() external payable;
-    function redeem(uint redeemTokens) external returns (uint);
-    function redeemUnderlying(uint redeemAmount) external returns (uint);
-    function borrow(uint borrowAmount) external returns (uint);
+
+    function redeem(uint256 redeemTokens) external returns (uint256);
+
+    function redeemUnderlying(uint256 redeemAmount) external returns (uint256);
+
+    function borrow(uint256 borrowAmount) external returns (uint256);
+
     function repayBorrow() external payable;
+
     function repayBorrowBehalf(address borrower) external payable;
+
     function liquidateBorrow(address borrower, address cTokenCollateral) external payable;
-    function balanceOfUnderlying(address account) external returns (uint);
+
+    function balanceOfUnderlying(address account) external returns (uint256);
 }
